client: modularized Step 3 component for ease of access
- renamed to step 2.5 with note on being intermediary
